Have the library handle challenge management (#35)
The examples so far have all used sessions to manage the active
challenges, but not all applications are stateful in this way - namely,
most APIs will not be session-based.

Instead, this creates a new `ChallengeManagerInterface` that handles
this for applications. For now there's a single implementation that's
still session-based, though (via #30 which I'm reworking) other
implementations will be provided (e.g. a cache pool).

The majority of the change here is updating examples and adding tests.
Note that this would be a BC break but since the library is still
pre-1.0 it's not a concern for practical purposes.