Expose user verified flag from parsed responses (#36)
While it's already possible to explicitly require the `userVerified`
flag during the registration and authentication ceremonies, it's
possible that a server may want to provide differing behavior when a
user was verified even if the ceremony expressed a preference of
`preferred` or even `discouraged`.

The common case here is to treat UV=1 as a two-factor auth, and UV=0 as
a single (possession) factor, and adjust permissions or a sign-in flow
accordingly. It's not strictly reliably accurate 100% of the time, and
this alone is insufficient to know _which_ factors were used, but it's a
close-enough assumption for the majority of use-cases. In any case, this
library imposes no requirements that servers even look at this, but it
provides the option.

Note that in theory the [`uvm`
extension](https://www.w3.org/TR/webauthn-2/#sctn-uvm-extension) should
provide similar information, MDN (as of writing) [doesn't
acknowledge](https://developer.mozilla.org/en-US/docs/Web/API/Web_Authentication_API/WebAuthn_extensions)
this value; I've had no success getting it to behave in any browser I've
tested. https://github.com/w3c/webauthn/issues/1386 /
https://github.com/w3c/webauthn/pull/1399 suggest it should continue to
exist, but in practice that seems not to be the case.