Codify all enums from the W3C spec (#47)
This imports (and converts to PHP) all of the remaining enums declared
(as of now) at https://w3c.github.io/webauthn/#idl-index. While the vast
majority are at present unused, support for most if not all will be
added in the future, particular for emitting JSON message formats to
simplify client code.