Restructure base64URL handling (#46)
The conversion is only really done in the context of a BinaryString, so
having the logic split out into a different class didn't make a lot of
sense. As a result, it's being consolidated. In addition to
organization, this should make completion of #41 easier, specifically
for emitting formats that match
`PublicKeyCredential.parse{Creation|Request}OptionsFromJSON()` methods.