Parse transports and pass them through to the credential (#55)
Pulled out of #53 - it's desirable to store the transports alongside the
credential in order to present future UI hints (note: it's not trusted
data). Since that PR was getting pretty large, that aspect is being
split out for readability.

Note: this maintains complete backwards compatibility with the original
wire format, but does add a new recommendation to start passing the
transports through.