Provide bypass for not-yet-supported attestation formats (#71)
By default, the library will require _any_ sort of "certain" trust path
during credential registration - this is actually slightly stricter than
the previous behavior, as `Packed` formats _can_ result in this path. To
compensate (and as a bonus, ease library development), this provides a
new flag that can be passed to the registration verification process
that _permits_ these uncertain paths to go through.

In effect, if this flag is set, any credential has the security
implications of the `none` format: you really know nothing about it, but
authentication still is cryptographically verified. In the common case
this means nothing, since (at minimum) Apple's implementation of
passkeys uses `fmt: none` and blocking them is likely undesirable for
most people.

Fixes #69.