Improve support for more complex challenge management (#75)
Splits the retrieval and management of challenges. 

From a good vibes perspective, this a) aligns the main interface name
with the functionality it's intended to provide, and b) improves the
interface segregation (hi, SOLID).

In terms of practical needs, this helps solve the (fairly rare) need of
challenges that are managed as part of a broader external system. This
could be a data model, microservice, or whatever. By splitting the
interfaces and only requiring that the newer `ChallengeLoaderInterface`
make it to the verification processes, this becomes both easier and more
testable.