Add 8.4 stable and 8.5-dev to test matrix (#88)
PHP 8.4 was released last week, so this updates the test matrix
accordingly. 8.1 can be reasonably dropped as well.

8.4 and 8.5 are now emitting a deprecation warning from the ASN.1
library. I intend to send a PR, but there's been no activity for quite
some time (understandable given the context of the library, but there's
also an open PR for other language-related improvements) so switching to
a more actively-maintained library may end up being a better choice.